requirements.txt
markdown==3.5.2
python-frontmatter==1.1.0
pathlib==1.0.1
openai==1.12.0
pinecone==7.0.2
python-dotenv==1.0.1
nltk==3.8.1
tqdm==4.66.2
tiktoken==0.9.0